Add routes for the transit gateway
Now non-local, non-COOL traffic goes to the NAT gateways or to the
internet gateway, while non-local, COOL traffic goes to the transit
gateway.

The routing configurations are a little more complicated now, so go
ahead and split out the public and private routing into separate
files.

We also need to add a new variable for the overall COOL CIDR block,
since that is required for the transit gateway routes.