Allow zone association with multiple VPCs
Private zones MUST ALWAYS be associated with a VPC, so I must use
inline vpc blocks when defining these zones.  Other VPCs may want to
associate with these zones as well (particularly the private zone - for DNS
purposes), though, and they will have to use a standalone
aws_route53_zone_association to do so.  Mixing inline and standalone
resources like this leads to a flapping of Terraform state, and the
only cure is to ignore changes to vpc.  Therefore I have to define
these lifecycle blocks.